move load_eye data function back to base
